src/site/requirements.txt
asgiref==3.8.1
blinker==1.4
requests==2.30
cryptography==3.4.8
charset-normalizer==3.3.2
channels==4.1.0
channels_redis==4.0
distlib==0.3.8
httplib2==0.20.2
Django==5.1.1
daphne==4.1.0
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg2-binary==2.9.10
sqlparse==0.5.1
typing_extensions==4.12.2
virtualenv==20.26.4
python-dotenv==1.0.1
gunicorn==23.0.0
Pillow==11
djangorestframework==3.15.2